+++ baselines/flwr_baselines/publications/fedprox_mnist/README.md
@@ -1,44 +1,60 @@
 # FedProx MNIST
 
-The following baseline replicates the experiments in *Federated Optimization in Heterogeneous Networks* (Li et al., 2018), which proposed the FedProx algorthim.
+> Note: If you use this baseline in your work, please remember to cite the original authors of the paper as well as the Flower paper.
 
-**Abstract:** Federated Learning is a distributed learning paradigm with two key challenges that differentiate it from traditional distributed optimization: (1) significant variability in terms of the systems characteristics on each device in the network (systems heterogeneity), and (2) non-identically distributed data across the network (statistical heterogeneity). In this work, we introduce a framework, FedProx, to tackle heterogeneity in federated networks. FedProx can be viewed as a generalization and re-parametrization of FedAvg, the current state-of-the-art method for federated learning. While this re-parameterization makes only minor modifications to the method itself, these modifications have important ramifications both in theory and in practice. Theoretically, we provide convergence guarantees for our framework when learning over data from non-identical distributions (statistical heterogeneity), and while adhering to device-level systems constraints by allowing each participating device to perform a variable amount of work (systems heterogeneity). Practically, we demonstrate that FedProx allows for more robust convergence than FedAvg across a suite of realistic federated datasets. In particular, in highly heterogeneous settings, FedProx demonstrates significantly more stable and accurate convergence behavior relative to FedAvg---improving absolute test accuracy by 22% on average.</i>
+**Paper:** https://arxiv.org/abs/1812.06127
 
-**Paper Authors:** 
+**Authors:** Tian Li, Anit Kumar Sahu, Manzil Zaheer, Maziar Sanjabi, Ameet Talwalkar and Virginia Smith.
 
-Tian Li, Anit Kumar Sahu, Manzil Zaheer, Maziar Sanjabi, Ameet Talwalkar and Virginia Smith.
+**Abstract:** Federated Learning is a distributed learning paradigm with two key challenges that differentiate it from traditional distributed optimization: (1) significant variability in terms of the systems characteristics on each device in the network (systems heterogeneity), and (2) non-identically distributed data across the network (statistical heterogeneity). In this work, we introduce a framework, FedProx, to tackle heterogeneity in federated networks. FedProx can be viewed as a generalization and re-parametrization of FedAvg, the current state-of-the-art method for federated learning. While this re-parameterization makes only minor modifications to the method itself, these modifications have important ramifications both in theory and in practice. Theoretically, we provide convergence guarantees for our framework when learning over data from non-identical distributions (statistical heterogeneity), and while adhering to device-level systems constraints by allowing each participating device to perform a variable amount of work (systems heterogeneity). Practically, we demonstrate that FedProx allows for more robust convergence than FedAvg across a suite of realistic federated datasets. In particular, in highly heterogeneous settings, FedProx demonstrates significantly more stable and accurate convergence behavior relative to FedAvg---improving absolute test accuracy by 22% on average.
 
-Note: If you use this implementation in your work, please remember to cite the original authors of the paper. 
 
-**[Link to paper.](https://arxiv.org/abs/1812.06127)**
+## About this baseline
+**What's implemented:** The code in this directory replicates the experiments in *Federated Optimization in Heterogeneous Networks* (Li et al., 2018) for MNIST, which proposed the FedProx algorthim. Concretely, it replicates the results for MNIST in Figure 1 and 7.
 
-## Training Setup
+**Datasets:** MNIST from PyTorch's Torchvision
 
-### Models
+**Hardware Setup:** These experiments were run on a desktop machine with 24 CPU threads. Any machine with 4 CPU cores of more would be able to run it in a reasonable amount of time.
 
-This baseline contains two models:
+**Contributors:** Charles Beauville and Javier Fernandez-Marques
 
+
+## Experimenatl Setup
+
+**Task:** Image classification
+
+**Model:** This directory implementes two models:
 * A logistic regression model used in the FedProx paper for MNIST (see `models/LogisticRegression`). This is the model used by default.
 * A two-layer CNN network as used in the FedAvg paper (see `models/Net`)
 
-### Training Paramaters
+**Dataset:** This baseline only includes the MNIST dataset. By default it will be partitioned into 1000 clients following a pathological split where each client has examples of two (out of ten) class labels. The number of examples in each client is derived by sampling from a powerlaw distribution. The settings are as follow:
 
-The following is a summary of the key hyperparameters most relevant for this baseline (FedProx+MNIST). These are the current defaults, but you can change them for your setting. For a complete list of hyperparameters and settings, please refer to `conf/config.yaml` (which runs FedProx) and `conf/fedavg.yaml` (which runs FedAvg dropping stragglers)
+| Dataset | #classes | #partitions | partitioning method | partition settings |
+| :------ | :---: | :---: | :---: | :---: |
+| MNIST | 10 | 1000 | pathological with power law | 2 classes per client |
 
-| Description | Value |
+**Training Hyperparameters:**
+The following table shows the main hyperparameters for this baseline with their default value (i.e. the value used if you run `python main.py` directly)
+
+| Description | Default Value |
 | ----------- | ----- |
 | total clients | 1000 |
 | clients per round | 10 |
-| data partition | power law (2 classes per client) |
+| number of rounds | 100 |
+| client resources | {'num_cpus': 2.0, 'num_gpus': 0.0 }|
+| data partition | pathological with power law (2 classes per client) |
 | optimizer | SGD with proximal term |
 | proximal mu | 1.0 |
 | stragglers_fraction | 0.9 |
 
-## Running experiments
+## Environment Setup
 
-The `conf/config.yaml` file containing all the tunable hyperparameters and the necessary variables. [Hydra](https://hydra.cc/docs/tutorials/) is used to manage configs. The outputs of each experiment as well as a log is created automatically by Hydra. The output directory will follow the structure: `outputs/<date>/<time>`.
+**TBD**
 
-To run FedPox as:
+## Running the Experiments
+
+Run this FedProx w/ MNIST baseline as:
+
 ```bash
 python main.py # this will run using the default settings in the `conf/config.yaml`
