+++ baselines/flwr_baselines/publications/fedprox_mnist/conf/fedavg.yaml
@@ -31,3 +31,6 @@ strategy:
   min_fit_clients: ${clients_per_round}
   min_available_clients: ${clients_per_round}
   min_evaluate_clients: 0
+  evaluate_metrics_aggregation_fn:
+    _target_: strategy.weighted_average
+    _partial_: true # we dont' want this function to be evaluated when instantiating the strategy, we treat it as a partial and evaluate it when the strategy actuallly calls the function (in aggregate_evaluate())
